app/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
cachetools==3.1.1
celery==4.3.0
chardet==3.0.4
Django==2.2.7
django-storages==1.8
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
gevent==1.4.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
greenlet==0.4.15
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
kombu==4.6.6
more-itertools==8.0.0
Pillow==6.2.1
protobuf==3.11.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.3.11
requests==2.22.0
rsa==4.0
sentry-sdk==0.13.2
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
